refactor(queue): keep shared contracts outside runtime owners
Move snapshot and block-reason types into the existing leaf module while preserving the public snapshot export. Register the narrow sparse-ring lint exception in the explicit inventory. Both changes address exact-head CI findings without changing emitted runtime code.
